requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework
# Para el modulo de balanza Lector de peso
requests
pyserial
python-dotenv
gunicorn==23.0.0
mysqlclient==2.2.6
whitenoise==6.8.2
